% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{nest}
\title{Nested function calls}
\usage{
nest(x, expr, n, m = 1, xname = "x")
}
\description{
Implementation of the Nest function from WolframAlpha language
}
\examples{
## fibonacci sequence
nest(c(1, 1), c(x[2], x[1] + x[2]), 10)
}

