% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_while}
\alias{nest_while}
\title{repeat until the condition is met}
\usage{
nest_while(x, expr, cond, m = 1, max = NA, n = 0, xname = "x")
}
\description{
repeat until the condition is met
}
\examples{
## collatz sequence
nest_while(1214, ifelse(x \%\% 2 == 0, x/2, 3*x+1), x != 1)
nest_while(c(3249, 2723, 4901), ifelse(x \%\% 2 == 0, x/2, ifelse(x == 1, 1, 3*x+1)), any(x != 1))
}

