% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{nest}
\title{Nested function calls}
\usage{
nest(expr, x, n, xname = "x")
}
\description{
Implementation of the Nest function from WolframAlpha language
}
\examples{
## fibonacci sequence
nest(c(x, x[length(x)] + x[length(x)-1]), c(1, 1), 5)
}

