% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{nest}
\title{Nested function calls}
\usage{
nest(expr, x, n, xname = "x")
}
\description{
Implementation of the Nest function from WolframAlpha language
}
\examples{
## fibonacci sequence
nest(c(x, x[length(x)] + x[length(x)-1]), c(1, 1), 5)

## creating a function
f <- nest((x+1)^2, NULL, 3)

## using different variable names
exp3 <- nest(exp(a), NULL, 3, xname = "a")

## parenthesis can make a difference
frac_right <- nest(1+1/(x), NULL, 4)
frac_wrong <- nest(1+1/x, NULL, 4)

}

